added v-usb library and done routing board
added v-usb library, but still need to change settings in usbconfig.h. also usb needs to use only one port for both d+ and d- lines (d+ can't be on pa0 and d- on pb1) so I had to change the usb data lines again. d+ is on PA0 with pcint0 and d- is on PA7. attiny85 pin 2 is now on PA6