Added Spice simulation files
added simulation.asc and zvp4105a.sub. I'm trying to make the simulation as accurate as possible to get a better understanding of what is happening since I don't have an oscilloscope to debug with. zvp4105a is currently not being used in the simulation, instead I'm using a pmos with a similar Rds